+++ src/node/behaviors/adopt.py
@@ -2,6 +2,7 @@ from __future__ import absolute_import
 from contextlib import contextmanager
 from node.interfaces import IMappingAdopt
 from node.interfaces import INode
+from node.interfaces import ISequenceAdopt
 from plumber import Behavior
 from plumber import plumb
 from zope.interface import implementer
@@ -33,15 +34,32 @@ def adopt_node(name, parent, value):
 class MappingAdopt(Behavior):
 
     @plumb
-    def __setitem__(next_, self, key, val):
-        with adopt_node(key, self, val):
-            next_(self, key, val)
+    def __setitem__(next_, self, key, value):
+        with adopt_node(key, self, value):
+            next_(self, key, value)
 
     @plumb
     def setdefault(next_, self, key, default=None):
-        # We reroute through __getitem__ and __setitem__, skipping next_
+        # Reroute through ``__getitem__`` and ``__setitem__``, skipping
+        # ``next_``.
         try:
             return self[key]
         except KeyError:
             self[key] = default
             return default
+
+
+@implementer(ISequenceAdopt)
+class SequenceAdopt(Behavior):
+
+    @plumb
+    def __setitem__(next_, self, index, value):
+        if type(index) is slice:
+            raise NotImplementedError('No slice support.')
+        with adopt_node(str(index), self, value):
+            next_(self, index, value)
+
+    @plumb
+    def insert(next_, self, index, value):
+        with adopt_node(str(index), self, value):
+            next_(self, index, value)